requirements.txt
tensorflow-gpu==2.10.1
tensorflow-intel==2.10.0
